refactor: send message correctly
