+++ src/app/validator-message/index.ts
@@ -8,4 +8,5 @@ export const validatorMessage = {
 	ROLE_MUST_BE_A_STRING: 'ROLE_MUST_BE_A_STRING',
 	INVALID_ROLE: 'INVALID_ROLE',
 	ROLE_IS_REQUIRED: 'ROLE_IS_REQUIRED',
+	INVALID_EMAIL: 'INVALID_EMAIL',
 };
